exception 처리, 영화 상세 페이지 수정, 리뷰 평점순/최신순 정렬, 모달 수정, 리뷰 등록, 조회, 삭제 구현

+++ src/main/java/com/example/vo/Movie.java
@@ -16,6 +16,6 @@ public class Movie {
 	private int no;
 	private int movieRatingNo;
 	private String movieName;
-	private int movieSalesRate;
+	private double movieSalesRate;
 	private int likeCount;
 }
